refactor: 💡 refactor hook auth login
move customHook to folder
